Move slices to separate folders.
