+++ src/app/slices/display/index.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { RootState } from "../../app/store";
+import { RootState } from "../../store";
 
 const applyTheme = (theme: "light" | "dark") => {
   document.documentElement.className = theme;
