+++ src/components/input/DrawerSearch.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import classNames from "classnames";
 import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { geocodeSearch } from "../../app/googleMaps";
-import { initialState, inputSetSearch, selectSearchQuery, setLatLng } from "./inputSlice";
-import { selectMapsLoading, selectMapsNoResults, selectMapsResult, setNoResults } from "./mapsSlice";
-import { selectTheme } from "../display/displaySlice";
+import { initialState, inputSetSearch, selectSearchQuery, setLatLng } from "../../app/slices/input";
+import { selectMapsLoading, selectMapsNoResults, selectMapsResult, setNoResults } from "../../app/slices/maps";
+import { selectTheme } from "../../app/slices/display";
 import { hasKey } from "../../util/functions";
 import { Button } from "@rmwc/button";
 import { Drawer, DrawerHeader, DrawerContent, DrawerTitle } from "@rmwc/drawer";
